+++ src/libs/satellite/src/rules/switch_memory.rs
